Fixed compatibility issue with AJAX Control Toolkit. (RM-4545)
git-svn-id: https://svn.re-motion.org/svn/Remotion/trunk@22099 cb0deee2-328e-b949-94ea-fb3139fab790

+++ Remotion/Web/Core/Res/HTML/SmartPage.js
@@ -208,7 +208,17 @@ function SmartPage_Context(
         delete this._scriptDisposes[updatePanelID];
       }
     }
-    this._destroyTree(updatePanelElement);
+    if (TypeUtility.IsDefined(this._destroyTree)) // .NET 3.5 AJAX library
+    {
+      this._destroyTree(updatePanelElement);
+    }
+    else if (TypeUtility.IsDefined(Sys.Application.disposeElement)) // .NET 4.0 AJAX library
+    {
+      Sys.Application.disposeElement(updatePanelElement, true);
+    }
+    else {
+      throw "Unsupported AJAX library detected."
+    }
     $(updatePanelElement).empty().append(rendering);
   }
